backend-python/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
requests